refactor user tab
